requirements.txt
aniso8601==9.0.1
black==23.3.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
PuLP==2.7.0
PyJWT==2.6.0
pymongo==4.3.3
pytz==2023.3
requests==2.31.0
six==1.16.0
urllib3==2.0.2
Werkzeug==2.3.3
